Connector/__MAIN__/requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
fastapi==0.98.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
